feat(content): 新增课程基本信息功能
- 添加 AddCourseDto 类用于新增课程参数接收
- 在 CourseBaseInfoController 中添加新增课程接口
- 在 CourseBaseInfoService接口中定义新增课程方法
- 实现 CourseBaseInfoServiceImpl 中的新增课程逻辑
- 创建 CourseBaseInfoVO 类用于返回课程基本信息- 更新 CourseMarket 类，将价格字段改为 BigDecimal 类型

+++ learn-online-content/learn-online-model/src/main/java/com/xia/content/model/po/CourseMarket.java
@@ -4,6 +4,7 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * <p>
@@ -31,12 +32,12 @@ public class CourseMarket implements Serializable {
     /**
      * 现价
      */
-    private Float price;
+    private BigDecimal price;
 
     /**
      * 原价
      */
-    private Float originalPrice;
+    private BigDecimal originalPrice;
 
     /**
      * 咨询qq
